Refactor function_record
